+++ packages/plugin-reanimated/loader/index.d.ts
@@ -1,0 +1,2 @@
+export * from '../dist/loader.js';
+export { default } from '../dist/loader.js';