+++ packages/repack/src/commands/webpack/start.ts
@@ -1,27 +1,28 @@
+// @ts-expect-error type-only import
 import type { Server } from '@callstack/repack-dev-server';
 import type { Config } from '@react-native-community/cli-types';
 import * as colorette from 'colorette';
 import type webpack from 'webpack';
 import packageJson from '../../../package.json';
-import { VERBOSE_ENV_KEY } from '../../env';
+import { VERBOSE_ENV_KEY } from '../../env.js';
 import {
   ConsoleReporter,
   FileReporter,
   type Reporter,
   composeReporters,
   makeLogEntryFromFastifyLog,
-} from '../../logging';
+} from '../../logging/index.js';
 import {
   getMimeType,
   getWebpackConfigFilePath,
   parseFileUrl,
   runAdbReverse,
   setupInteractions,
-} from '../common';
-import { DEFAULT_HOSTNAME, DEFAULT_PORT } from '../consts';
-import type { StartArguments, StartCliOptions } from '../types';
-import { Compiler } from './Compiler';
-import type { HMRMessageBody } from './types';
+} from '../common/index.js';
+import { DEFAULT_HOSTNAME, DEFAULT_PORT } from '../consts.js';
+import type { StartArguments, StartCliOptions } from '../types.js';
+import { Compiler } from './Compiler.js';
+import type { HMRMessageBody } from './types.js';
 
 /**
  * Start command for React Native Community CLI.
@@ -41,6 +42,13 @@ export async function start(_: string[], config: Config, args: StartArguments) {
     args.config ?? args.webpackConfig
   );
   const { reversePort, ...restArgs } = args;
+
+  const serverProtocol = args.https ? 'https' : 'http';
+  const serverHost = args.host || DEFAULT_HOSTNAME;
+  const serverPort = args.port ?? DEFAULT_PORT;
+  const serverURL = `${serverProtocol}://${serverHost}:${serverPort}`;
+  const showHttpRequests = args.verbose || args.logRequests;
+
   const cliOptions: StartCliOptions = {
     config: {
       root: config.root,
@@ -49,7 +57,9 @@ export async function start(_: string[], config: Config, args: StartArguments) {
       reactNativePath: config.reactNativePath,
     },
     command: 'start',
-    arguments: { start: { ...restArgs } },
+    arguments: {
+      start: { ...restArgs, host: serverHost, port: serverPort },
+    },
   };
 
   if (args.platform && !cliOptions.config.platforms.includes(args.platform)) {
@@ -73,11 +83,6 @@ export async function start(_: string[], config: Config, args: StartArguments) {
   );
 
   const compiler = new Compiler(cliOptions, reporter);
-
-  const serverHost = args.host || DEFAULT_HOSTNAME;
-  const serverPort = args.port ?? DEFAULT_PORT;
-  const serverURL = `${args.https === true ? 'https' : 'http'}://${serverHost}:${serverPort}`;
-  const showHttpRequests = args.verbose || args.logRequests;
 
   const { createServer } = await import('@callstack/repack-dev-server');
   const { start, stop } = await createServer({
@@ -108,13 +113,20 @@ export async function start(_: string[], config: Config, args: StartArguments) {
                 method: 'POST',
               });
             },
+            onAdbReverse() {
+              void runAdbReverse({
+                port: serverPort,
+                logger: ctx.log,
+                verbose: true,
+              });
+            },
           },
           { logger: ctx.log }
         );
       }
 
-      if (reversePort && args.port) {
-        void runAdbReverse(args.port, ctx.log);
+      if (reversePort) {
+        void runAdbReverse({ logger: ctx.log, port: serverPort, wait: true });
       }
 
       const lastStats: Record<string, webpack.StatsCompilation> = {};
@@ -122,7 +134,7 @@ export async function start(_: string[], config: Config, args: StartArguments) {
       compiler.on('watchRun', ({ platform }) => {
         ctx.notifyBuildStart(platform);
         if (platform === 'android') {
-          void runAdbReverse(args.port ?? DEFAULT_PORT, ctx.log);
+          void runAdbReverse({ port: serverPort, logger: ctx.log });
         }
       });