+++ packages/repack/src/plugins/ModuleFederationPluginV2.ts
@@ -1,6 +1,6 @@
 import type { moduleFederationPlugin as MF } from '@module-federation/sdk';
 import type { Compiler, RspackPluginInstance } from '@rspack/core';
-import { isRspackCompiler } from './utils/isRspackCompiler';
+import { isRspackCompiler } from './utils/isRspackCompiler.js';
 
 /**
  * {@link ModuleFederationPlugin} configuration options.
@@ -11,6 +11,15 @@ import { isRspackCompiler } from './utils/isRspackCompiler';
  */
 export interface ModuleFederationPluginV2Config
   extends MF.ModuleFederationPluginOptions {
+  /**
+   *  List of default runtime plugins for Federation Runtime.
+   *  Useful if you want to modify or disable behaviour of runtime plugins.
+   *
+   *  Defaults to an array containing:
+   *    - '@callstack/repack/mf/core-plugin
+   *    - '@callstack/repack/mf/resolver-plugin
+   */
+  defaultRuntimePlugins?: string[];
   /** Enable or disable adding React Native deep imports to shared dependencies. Defaults to true */
   reactNativeDeepImports?: boolean;
 }
@@ -81,13 +90,19 @@ export interface ModuleFederationPluginV2Config
  * @category Webpack Plugin
  */
 export class ModuleFederationPluginV2 implements RspackPluginInstance {
-  private config: MF.ModuleFederationPluginOptions;
+  public config: MF.ModuleFederationPluginOptions;
   private deepImports: boolean;
+  private defaultRuntimePlugins: string[];
 
   constructor(pluginConfig: ModuleFederationPluginV2Config) {
-    const { reactNativeDeepImports, ...config } = pluginConfig;
+    const { defaultRuntimePlugins, reactNativeDeepImports, ...config } =
+      pluginConfig;
     this.config = config;
     this.deepImports = reactNativeDeepImports ?? true;
+    this.defaultRuntimePlugins = defaultRuntimePlugins ?? [
+      '@callstack/repack/mf/core-plugin',
+      '@callstack/repack/mf/resolver-plugin',
+    ];
   }
 
   private ensureModuleFederationPackageInstalled(context: string) {
@@ -105,10 +120,6 @@ export class ModuleFederationPluginV2 implements RspackPluginInstance {
     context: string,
     runtimePlugins: string[] | undefined = []
   ) {
-    const repackRuntimePlugin = require.resolve(
-      '../modules/FederationRuntimePlugin'
-    );
-
     const plugins = runtimePlugins
       .map((pluginPath) => {
         try {
@@ -121,11 +132,14 @@ export class ModuleFederationPluginV2 implements RspackPluginInstance {
       })
       .filter((pluginPath) => !!pluginPath) as string[];
 
-    if (!plugins.includes(repackRuntimePlugin)) {
-      return [repackRuntimePlugin, ...runtimePlugins];
+    for (const plugin of this.defaultRuntimePlugins) {
+      const pluginPath = require.resolve(plugin);
+      if (!plugins.includes(pluginPath)) {
+        plugins.unshift(pluginPath);
+      }
     }
 
-    return runtimePlugins;
+    return plugins;
   }
 
   private getModuleFederationPlugin(compiler: Compiler) {
@@ -219,15 +233,35 @@ export class ModuleFederationPluginV2 implements RspackPluginInstance {
     return adjustedSharedDependencies;
   }
 
-  apply(compiler: Compiler) {
-    this.ensureModuleFederationPackageInstalled(compiler.context);
-
+  private setupIgnoredWarnings(compiler: Compiler) {
     // MF2 produces warning about not supporting async await
     // we can silence this warning since it works just fine
     compiler.options.ignoreWarnings = compiler.options.ignoreWarnings ?? [];
     compiler.options.ignoreWarnings.push(
       (warning) => warning.name === 'EnvironmentNotSupportAsyncWarning'
     );
+    // MF2 produces warning about dynamic import in loadEsmEntry but it's not relevant
+    // in RN env since we override the loadEntry logic through a hook
+    // https://github.com/module-federation/core/blob/fa7a0bd20eb64eccd6648fea340c6078a2268e39/packages/runtime/src/utils/load.ts#L28-L37
+    compiler.options.ignoreWarnings.push((warning) => {
+      // @ts-expect-error moduleDescriptor is present in the warning object
+      const modulePath = warning.moduleDescriptor.name;
+      const moduleName = '@module-federation/runtime/dist/index.cjs.js';
+      const isMF2Runtime = modulePath.endsWith(moduleName);
+      const requestExpressionWarning =
+        /Critical dependency: the request of a dependency is an expression/;
+
+      if (isMF2Runtime && requestExpressionWarning.test(warning.message)) {
+        return true;
+      }
+
+      return false;
+    });
+  }
+
+  apply(compiler: Compiler) {
+    this.ensureModuleFederationPackageInstalled(compiler.context);
+    this.setupIgnoredWarnings(compiler);
 
     const ModuleFederationPlugin = this.getModuleFederationPlugin(compiler);