+++ packages/repack/src/plugins/RepackTargetPlugin/RepackTargetPlugin.ts
@@ -61,6 +61,11 @@ export class RepackTargetPlugin implements RspackPluginInstance {
     compiler.options.output.chunkFormat = 'array-push';
     compiler.options.output.globalObject = globalObject;
 
+    // Disable built-in strict module error handling
+    // this is handled through an interceptor in the
+    // init module added to __webpack_require__.i array
+    compiler.options.output.strictModuleErrorHandling = false;
+
     // Normalize global object.
     new compiler.webpack.BannerPlugin({
       raw: true,
@@ -75,7 +80,7 @@ export class RepackTargetPlugin implements RspackPluginInstance {
     new compiler.webpack.NormalModuleReplacementPlugin(
       /react-native.*?([/\\]+)Libraries([/\\]+)Utilities([/\\]+)HMRClient\.js$/,
       (resource) => {
-        const request = require.resolve('../../modules/DevServerClient');
+        const request = require.resolve('../../modules/DevServerClient.js');
         const context = path.dirname(request);
         resource.request = request;
         resource.context = context;
@@ -89,7 +94,7 @@ export class RepackTargetPlugin implements RspackPluginInstance {
     // ReactNativeTypes.js is flow type only module
     new compiler.webpack.NormalModuleReplacementPlugin(
       /react-native.*?([/\\]+)Libraries[/\\]Renderer[/\\]shims[/\\]ReactNativeTypes\.js$/,
-      require.resolve('../../modules/EmptyModule')
+      require.resolve('../../modules/EmptyModule.js')
     ).apply(compiler);
 
     compiler.hooks.thisCompilation.tap('RepackTargetPlugin', (compilation) => {
@@ -106,7 +111,7 @@ export class RepackTargetPlugin implements RspackPluginInstance {
             const loadScriptGlobal = compiler.webpack.RuntimeGlobals.loadScript;
             const loadScriptRuntimeModule = Template.asString([
               Template.getFunctionContent(
-                require('./implementation/loadScript')
+                require('./implementation/loadScript.js')
               )
                 .replaceAll('$loadScript$', loadScriptGlobal)
                 .replaceAll('$caller$', `'${chunk.id?.toString()}'`),
@@ -114,7 +119,7 @@ export class RepackTargetPlugin implements RspackPluginInstance {
 
             const initRuntimeModule = Template.asString([
               '// Repack runtime initialization logic',
-              Template.getFunctionContent(require('./implementation/init'))
+              Template.getFunctionContent(require('./implementation/init.js'))
                 .replaceAll('$globalObject$', globalObject)
                 .replaceAll('$hmrEnabled$', `${this.config?.hmr ?? false}`),
             ]);