+++ packages/repack/src/plugins/__tests__/ModuleFederationPluginV2.test.ts
@@ -1,6 +1,6 @@
 import { ModuleFederationPlugin as MFPluginRspack } from '@module-federation/enhanced/rspack';
 import type { Compiler } from '@rspack/core';
-import { ModuleFederationPluginV2 } from '../ModuleFederationPluginV2';
+import { ModuleFederationPluginV2 } from '../ModuleFederationPluginV2.js';
 
 jest.mock('@module-federation/enhanced/rspack');
 
@@ -16,8 +16,9 @@ const mockPlugin = MFPluginRspack as unknown as jest.Mock<
   typeof MFPluginRspack
 >;
 
-const runtimePluginPath = require.resolve(
-  '../../modules/FederationRuntimePlugin'
+const corePluginPath = require.resolve('@callstack/repack/mf/core-plugin');
+const resolverPluginPath = require.resolve(
+  '@callstack/repack/mf/resolver-plugin'
 );
 
 describe('ModuleFederationPlugin', () => {
@@ -136,22 +137,24 @@ describe('ModuleFederationPlugin', () => {
     expect(config.shared['@react-native/'].eager).toBe(false);
   });
 
-  it('should add FederationRuntimePlugin to runtime plugins', () => {
+  it('should add CorePlugin & ResolverPlugin to runtime plugins by default', () => {
     new ModuleFederationPluginV2({ name: 'test' }).apply(mockCompiler);
 
     const config = mockPlugin.mock.calls[0][0];
-    expect(config.runtimePlugins).toContain(runtimePluginPath);
+    expect(config.runtimePlugins).toContain(corePluginPath);
+    expect(config.runtimePlugins).toContain(resolverPluginPath);
   });
 
-  it('should not add FederationRuntimePlugin to runtime plugins when already present', () => {
+  it('should not add duplicate default runtime plugins when already present', () => {
     new ModuleFederationPluginV2({
       name: 'test',
-      runtimePlugins: [runtimePluginPath],
+      runtimePlugins: [corePluginPath, resolverPluginPath],
     }).apply(mockCompiler);
 
     const config = mockPlugin.mock.calls[0][0];
-    expect(config.runtimePlugins).toContain(runtimePluginPath);
-    expect(config.runtimePlugins).toHaveLength(1);
+    expect(config.runtimePlugins).toContain(corePluginPath);
+    expect(config.runtimePlugins).toContain(resolverPluginPath);
+    expect(config.runtimePlugins).toHaveLength(2);
   });
 
   it('should use loaded-first as default shareStrategy', () => {